+++ changelog.md
@@ -4,6 +4,7 @@
 * Show "amount is more than allowed for online booking" message directly underneath the corresponding input instead of at the end of the form.
 * Add `recras-input-invalid` class to invalid amount inputs
 * Use minimum amount instead of number of people for "minimum amount" calculation
+* Fix book button being enabled despite "I agree" checkbox not being checked
 
 ## 0.10.3 (2018-11-20)
 * Fix standalone contact forms